pyproject.toml
[project]
name = "function_server"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "gchuyun", email = "gchuyun@qq.com" }
]
dependencies = [
    "fastapi>=0.110.1",
    "uvicorn>=0.29.0",
    "pip>=24.0",
    "langchain>=0.1.16",
    "langchain-openai>=0.1.3",
    "pydantic>=2.7.0",
    "beautifulsoup4>=4.12.3",
    "openai-function-calling>=2.1.0",
    "nb-log>=12.6",
    "loguru>=0.7.2",
    "search-engines-scraper-tasos>=0.6",
    "environs>=11.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/function_server"]

[project.scripts]
function_server = 'function_server.main:main'
websearch = 'function_server.tools.websearch:main'

[tool.rye.scripts]
dev = { cmd = "uvicorn function_server.main:app --host '0.0.0.0' --reload", env = { REQUESTS_CA_BUNDLE = "", LOG_LEVEL = "DEBUG" } }
